pyproject.toml
[project]
name = "crewai-playground"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Olaf", email = "olaf@pobox.com"},
]
dependencies = [
    "crewai>=0.27.0",
    "python-dotenv>=1.0.0",
    "duckduckgo-search>=5.1.0",
    "openai>=1.12.0",
    "unstructured>=0.12.4",
    "pyowm>=3.3.0",
    "tools>=0.1.9",
    "wikipedia>=1.4.0",
    "yfinance>=0.2.36",
    "sec-api>=1.0.17",
    "tiktoken>=0.5.2",
    "faiss-cpu>=1.7.4",
    "python-box>=7.1.1",
    "langchain-mistralai>=0.0.4",
    "langchain-community>=0.0.19",
    "langchain-experimental>=0.0.49",
    "e2b>=0.14.2",
    "aifs>=0.0.7",
    "chromadb>=0.4.22",
    "ollama>=0.1.6",
    "langchain-google-genai>=0.0.9",
    "pygame>=2.5.2",
    "crewai-tools>=0.0.15",
    "langchain-anthropic>=0.1.4",
]
requires-python = ">=3.10,<3.12"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

[tool.pdm.dev-dependencies]
test = [
    "pytest>=8.0.0",
]
